+++ shared_data.c
@@ -5,8 +5,8 @@
 // Initialises the shared_data variables
 void initialise_shared_data_2d(
     const int global_nx, const int global_ny, const int local_nx, 
-    const int local_ny, const int x_off, const int y_off, const int mesh_width, 
-    const int mesh_height, const char* problem_def_filename, const double* edgex, 
+    const int local_ny, const int x_off, const int y_off, const double mesh_width, 
+    const double mesh_height, const char* problem_def_filename, const double* edgex, 
     const double* edgey, SharedData* shared_data) 
 {
   const int ndims = 2;
